Travel Guide Filtering AI - version 1
About: Automating the process of filtering invalid travel guides with NLP.

Data mining: Collected valid and invalid samples manually through the use of ChatGPT, TripAdvisor, MyIreland, etc. Fully aware that the data that we have is very limited (only 600 data) in total, and the invalid data is biased towards hate speech since the distribution is not even. However, we found that in general, if it's a "perfectly" valid travel guide or a "perfectly" invalid text, the AI will be able to detect it. TODO: need to source more data, especially on the invalid spectrum, add more types of samples.

Data processing: Currently no special data processing done. The only data processing we do is tokenization (which is a must for NLP) and truncating text that has a max_len value more than the model's capacity.

Model: We use BERT as our model. We're using BERT since it is an encoder-transformer, which was trained bidirectionally, i.e., the context includes both words on the left and on the right.

Labels/Classes: We're currently only classifying the text into 3 different classes: valid, invalid - harassment, and invalid - others. We hope to add more classes in the future as we add more data.

Accuracy: Since we have a very limited number of data, our validation dataset percentage is only 0.1 (10%). Because of this, it was easy for us to get 100% on the accuracy. However, had the data been more sparse, it would've yielded a different result.

Server: Created a simple Flask app to serve requests to the model.